Move ACL manager to its own packaghe
