+++ pom.xml
@@ -10,6 +10,7 @@
     <module>pass-authz-integration</module>
     <module>pass-user-service</module>
     <module>pass-authz-filters</module>
+    <module>pass-authz-acl</module>
   </modules>
 
   <properties>
